Done some small refactoring to get all the middleware consepts and naming good
