+++ service_example/conf/routes
@@ -7,6 +7,6 @@
 #GET     /swagger.json               controllers.ApiHelpController.getResources
 
 # Setup endpoints
-GET     /v1                         no.uio.musit.microservices.example.controllers.Microservice.list
-POST    /v1                         no.uio.musit.microservices.example.controllers.Microservice.add
+GET     /v1                         no.uio.musit.microservice.example.resource.ExampleResource_V1.list
+POST    /v1                         no.uio.musit.microservice.example.resource.ExampleResource_V1.add
 #GET    /api-docs/v1                controllers.ApiHelpController.getResource(path = "/v1")
